// frontend/src/api/cv/models/ResumeRecord.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ResumeRecord = {
    id: string;
    user_id: string;
    candidate_name: string;
    candidate_age?: number | null;
    experience_years: string;
    file_url: string;
    analysis: string;
    created_at: string;
    updated_at: string;
};