// frontend/src/api/cv/models/UploadDatabaseResponse.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ResumeRecord } from './ResumeRecord';
export type UploadDatabaseResponse = {
    processed_count: number;
    successful_count: number;
    failed_count: number;
    resumes: Array<ResumeRecord>;
};